requirements.txt
spotipy
pandas
matplotlib
mysql-connector-python
python-dotenv